# openapi.yaml
...
openapi: 3.0.3
info:
  title: PrepChef API
  description: Commercial kitchen marketplace API
  version: 1.0.0
  contact:
    email: api@prepchef.com
  license:
    name: MIT
servers:
  - url: https://api.prepchef.com/v1
    description: Production
  - url: https://staging-api.prepchef.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
...
  /api/v1/admin/kitchens/pending:
    get:
      tags: [Admin]
      summary: List kitchens awaiting moderation
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Cursor from the previous moderation page.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Maximum number of kitchens to return.
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive search across kitchen and host name fields.
        - name: owner_email
          in: query
          required: false
          schema:
            type: string
        - name: certification_status
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Moderation queue page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminKitchenPage'
        '422':
          description: Invalid cursor supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/certifications/pending:
    get:
      tags: [Admin]
      summary: List pending certification documents
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Cursor from the previous certification page.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: document_type
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Certification review queue page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCertificationPage'
        '422':
          description: Invalid cursor supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/users:
    get:
      tags: [Admin]
      summary: List user accounts
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: role
          in: query
          required: false
          schema:
            type: string
        - name: include_suspended
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User directory page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserPage'
        '422':
          description: Invalid cursor supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        target:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
    ErrorResponse:
      type: object
      required: [request_id, error]
      properties:
        request_id:
          type: string
        error:
          $ref: '#/components/schemas/ErrorDetail'
    CursorPageMeta:
      type: object
      required: [limit, has_more]
      properties:
        cursor:
          type: string
          nullable: true
        next_cursor:
          type: string
          nullable: true
        limit:
          type: integer
          minimum: 1
          maximum: 100
        has_more:
          type: boolean
...
    AdminKitchenSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        owner_id:
          type: string
          format: uuid
        owner_email:
          type: string
        owner_name:
          type: string
        location:
          type: string
        submitted_at:
          type: string
          format: date-time
        moderation_status:
          type: string
        certification_status:
          type: string
      required: [id, name, owner_id, owner_email, submitted_at, moderation_status, certification_status]
    AdminKitchenPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminKitchenSummary'
        pagination:
          $ref: '#/components/schemas/CursorPageMeta'
    AdminCertificationSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        kitchen_id:
          type: string
          format: uuid
        kitchen_name:
          type: string
        document_type:
          type: string
        status:
          type: string
        submitted_at:
          type: string
          format: date-time
      required: [id, kitchen_id, document_type, status, submitted_at]
    AdminCertificationPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminCertificationSummary'
        pagination:
          $ref: '#/components/schemas/CursorPageMeta'
    AdminUserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        is_suspended:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, email, full_name, role, is_active, is_suspended, created_at]
    AdminUserPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserSummary'
        pagination:
          $ref: '#/components/schemas/CursorPageMeta'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []